Make source an optional thing to specify
sourceType is no longer required for initializing a SignalFxReport.
It is also possible to set the name of the "source" dimension with
via App.config as sourceDimension attribute or using the
.WithSourceDimension(..) method on a SignalFxReportBuilder.

For backward compatility if sourceDimension is not specified and
as source type is used, the old "sf_source" sourceDimension will
be used.