Change the default host dimension
Change the default host dimension used by Metrics.NET.SignalFx.dll from
the legacy "sf_source" to "host" to be consistent with other platforms
and so that Windows hosts will show up on the "Hosts" Page.

Retain the legacy dimension as an unused-by-default static string with
appropriate explanation so that it is not lost to the annals of time for
people wanting to go backwards for some unforseen reason.

Remove an unused copy of the legacy source dimension from the
SignalFxReport object.

Bump the version of Metrics.NET.SignalFx.dll to 2.5.0.0.